api/pyproject.toml
[tool.poetry]
name = "stock-market"
version = "0.1.0"
description = ""
authors = ["Matheus Takata <matheus.takata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
beautifulsoup4 = "^4.12.3"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
